Extract parsed city object projection
